Explicitly build local images for linux/amd64 for now
Tried building images on an M1 Mac for the first time and ran into the fact that not every upstream image exists for linux/arm64 (which is handled by the clojure manifest for the official images). So for now I'm just explicitly building on linux/amd64 locally (which is supported via qemu on M1 Macs w/ buildx). Should make it smarter at some point.